requirements.txt
streamlit>=1.28.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-core
langchain-chroma>=0.1.0
openai>=1.10.0
chromadb>=0.4.22,<0.5
pandas>=2.0.0
openpyxl>=3.1.0
numpy==1.26.4
python-dotenv>=1.0.0
tiktoken>=0.5.0
requests>=2.31.0
markdown>=3.4.0
beautifulsoup4>=4.11.0
transformers==4.55.2
